// src/api/generated/schemas/ongoingFestivalItemDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */

export interface OngoingFestivalItemDto {
  /** @nullable */
  imageUrl: string | null;
  id: string;
  name: string;
  regionLabel?: string;
  daysToStart?: number;
  eventStartDate?: string;
  eventEndDate?: string;
}